# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "rfc3339.date — Time API",
    "version": "1.0.0",
    "summary": "Strict RFC3339 time endpoints + validation + conversions (incl. exotic time standards)",
    "description": "A precise, developer-centric time API. Defaults to canonical RFC3339 UTC for /now, supports IANA time zones, strict RFC3339 validation, ISO8601 validation, and conversions across common + exotic representations (NTP, GPS, TAI, JD/MJD, Excel, HTTP-date, RFC5322). Plain text is first-class; JSON is optional via Accept header or ?json=1.\n",
    "license": {
      "name": "UNLICENSED",
      "identifier": "NOASSERTION"
    }
  },
  "servers": [
    {
      "url": "https://rfc3339.date",
      "description": "Production"
    },
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local dev (wrangler)"
    }
  ],
...
  "security": [],
  "paths": {
...
    "/tz": {
      "get": {
        "operationId": "listTimezones",
        "tags": [
          "Timezones"
        ],
        "summary": "Time zone search / listing",
        "description": "Returns a list of IANA time zone IDs matching q (substring match). If q is omitted, returns a curated list (or an implementation-defined subset).\n",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Zone matches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "zones"
                  ],
                  "properties": {
                    "zones": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "exampleCity": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid input, ambiguity, or unsupported combination)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tz/{zone}/offset": {
      "get": {
        "operationId": "getTimezoneOffset",
        "tags": [
          "Timezones"
        ],
        "summary": "Offset/DST information for a zone at a given instant",
        "parameters": [
          {
            "name": "zone",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "IANA time zone ID (URL-encoded), e.g. Europe%2FBerlin"
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "RFC3339 instant to evaluate; default is now."
          },
          {
            "name": "json",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ],
              "default": 0
            },
            "description": "If 1, forces JSON output regardless of Accept header."
          }
        ],
        "responses": {
          "200": {
            "description": "Offset and DST status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "zone",
                    "at",
                    "offset",
                    "dst"
                  ],
                  "properties": {
                    "zone": {
                      "type": "string"
                    },
                    "at": {
                      "type": "string"
                    },
                    "offset": {
                      "type": "string"
                    },
                    "dst": {
                      "type": "boolean"
                    },
                    "abbreviation": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "description": "Best-effort (may be null; abbreviations are not stable/portable)."
                    }
                  }
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Single-line summary: <offset> <dstFlag>"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid input, ambiguity, or unsupported combination)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found (unknown time zone or resource)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tz/{zone}/transitions": {
      "get": {
        "operationId": "listTimezoneTransitions",
        "tags": [
          "Timezones"
        ],
        "summary": "DST/offset transitions for a zone in a time range",
        "description": "Returns boundaries where the numeric offset changes within the requested range. Implementation can compute transitions by scanning offsets and refining boundary instants.\n",
        "parameters": [
          {
            "name": "zone",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "RFC3339 instant (inclusive)"
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "RFC3339 instant (exclusive)"
          },
          {
            "name": "granularity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "day",
                "hour",
                "minute"
              ],
              "default": "hour"
            },
            "description": "Scan step used to detect transitions before boundary refinement."
          },
          {
            "name": "refine",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "s",
                "ms"
              ],
              "default": "s"
            },
            "description": "Refinement resolution for transition instant."
          }
        ],
        "responses": {
          "200": {
            "description": "Transition list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "zone",
                    "start",
                    "end",
                    "transitions"
                  ],
                  "properties": {
                    "zone": {
                      "type": "string"
                    },
                    "start": {
                      "type": "string"
                    },
                    "end": {
                      "type": "string"
                    },
                    "transitions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "at",
                          "offsetBefore",
                          "offsetAfter"
                        ],
                        "properties": {
                          "at": {
                            "type": "string",
                            "description": "RFC3339 instant where the new offset takes effect (best-effort resolution)"
                          },
                          "offsetBefore": {
                            "type": "string"
                          },
                          "offsetAfter": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid input, ambiguity, or unsupported combination)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found (unknown time zone or resource)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "message"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "at": {
                            "type": "integer",
                            "description": "Character index (if applicable)"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}